+++ drivers/gpu/nvgpu/common/pmu/pmgr/pwrmonitor.c
@@ -145,7 +145,7 @@ done:
 }
 
 static struct boardobj *construct_pwr_topology(struct gk20a *g,
-				void *pargs, u16 pargs_size, u8 type)
+				void *pargs, size_t pargs_size, u8 type)
 {
 	struct boardobj *board_obj_ptr = NULL;
 	int status;
@@ -190,7 +190,7 @@ static int devinit_get_pwr_topology_table(struct gk20a *g,
 	struct pwr_topology_2x_entry pwr_topology_table_entry = { 0 };
 	u32 index;
 	u32 obj_index = 0;
-	u16 pwr_topology_size;
+	size_t pwr_topology_size;
 	union {
 		struct boardobj boardobj;
 		struct pwr_channel pwrchannel;
